cmake: Move grit helpers to libgbaline folder
A project that uses libgbaline should just be able to do something like
``add_subdirectory(libgbaline)`` and get the ability to call the grit
helpers. Right now, they are part of the gbaline folder, so a project
would have to copy the file to its own folder, which is wasteful.