library: build: Fix shared library creation in MSVC
In MSVC it is needed to specify the symbols that are exported from the
DLL in order for the program to see them. In GCC, this isn't needed.

In order to make the code compatible with both, a new macro has been
added (EXPORT_API) that has to be used by all exported functions of the
shared library.