library: Make per-frame execution more accurate
Previously, the script was running in parallel and it was impossible to
make sure that the game would run a specified number of frames before it
stopped (for a screenshot, for example).

Now, the run_frames_and_pause() command is capable of resuming the
execution and continue() isn't needed. Also, the game starts paused and
only run_frames_and_pause() or continue() can start it.

If the game is paused because of run_frames_and_pause(), it is possible
to call run_frames_and_pause() right away, which will make sure that the
number of frames executed at each step is the exact one that the user
asked for.