examples: Improve timer tests
Now they always generate a reference image in case they fail so that it
is possible to see the values that caused the test to fail. This could
be useful to adjust the tolerances of the test.

It also disables the timers as soon as the main test loop ends so that
the values don't change between the end of the test and when the ranges
are checked.