examples: cmake: Add support for unit tests in emulator
Until now, all tests were done in the binaries generated for the PC.
This is useful to test the SDL2 port of the library, but it isn't
perfect and it's easy to introduce errors that only affect the GBA port.

This commit adds custom targets to the CMake build system to build the
library for GBA, and to build all the tests for GBA (that depend on the
library being built first).

Also, increase the wait time in the framebuffer unit tests. The PC
version passes because the PC has unlimited computing power compared to
the GBA. In the GBA it takes a while to draw the full framebuffer as in
the examples.