template: Decouple from main build system
This is meant to be an example of how to integrate the library in a
project. It doesn't make sense to rely on the same build system of the
library as any other project won't be doing it this way.

This breaks the template build, which will be fixed in successive
commits.