examples: Remove `-sdl2` from all test filenames
This is now optional. For example, if file "test-sdl2.lua" is found, it
has priority over "test.lua". The same goes for "test-gba.lua", and for
the reference png files. This means that most tests can share the script
and reference images, and only the ones that need special treatment in
one platform need to have special files per platform.