library: unittests: Fix FP_Sin(x) when x = pi/2
This is an important point of the function that needs to be 100%
accurate. The reason why this angle wasn't giving the correct result is
that the last shift of the function was:

    0xFFFFFFC000 >> 24 = 0xFFFF.FFC

Unfortunately, this doesn't round up correctly, it just clamps. By
adding 0.5 << 24 before the shift, the result rounds up correctly.

The unit test now expects all multiples of pi/2 to be 100% accurate to
pass the test.