library: Improve arguments of TM_TimerStartUse()
Instead of using hardware definitions for the prescaler value, use
integers. This is easier to understand from a user's perspective, and
shorter to write.

Fix examples that relied on the old interface.

Update reference WAV for GBA test in sound example. The timing has changed
a bit because of the extra checks in the timer functions, so the reference
is no longer valid.