cmake: Add labels to tests
Each test is now assigned a label corresponding to the name of the
parent folder. For example, for the test ``examples/bios/math_arctan``,
the test name is ``math_arctan``, and the group name is ``bios``.

It is possible to only test a group of tests with ``ctest -L bios``.