libsysgba: Stop building it as a library
Building it as a library had problems because static libraries have to
be passed in the right order to the linker.

The macros of libugba and libsysgba have been modified so that now the
files are added to the project rather than building libsysgba and
linking the project to libsysgba.