library: Remove packed attribute from structs
This also fixes the issue mentioned in commit dcef6bdb51d1 ("docs: Add
known bug to list of bugs").

In short, packing structs make all the fields be aligned to 1 byte. This
is ok if the struct is in regular RAM, but it's a problem if the struct
is in memory that can't be modified with byte accesses.

The struct was packed, so some builds generated writes of one byte,
which is a problem in the real GBA because only 16 and 32-bit accesses
are allowed in OAM and VRAM.

Note that enabling optimizations gets rid of this issue even without the
fix in this commit.

Thanks to @LunarLambda for the help.

Source:

    https://gcc.gnu.org/onlinedocs/gcc-8.4.0/gcc/Structure-Layout-Pragmas.html

    For compatibility with Microsoft Windows compilers, GCC supports a
    set of #pragma directives that change the maximum alignment of
    members of structures (other than zero-width bit-fields), unions,
    and classes subsequently defined. The n value below always is
    required to be a small power of two and specifies the new alignment
    in bytes.